+++ Yinyang.Utilities/ExtensionString.cs
@@ -1,4 +1,7 @@
 using System;
+using System.IO;
+using System.Linq;
+using System.Text;
 using System.Text.RegularExpressions;
 
 namespace Yinyang.Utilities
@@ -178,6 +181,63 @@ namespace Yinyang.Utilities
             }
 
             return Regex.IsMatch(str, pattern);
+        }
+
+        /// <summary>
+        ///    Removes all line breaks (both \r and \n) from the given string.
+        /// </summary>
+        /// <remarks>文字列からすべての改行コード（\r および \n）を削除します。</remarks>
+        /// <param name="str">The input string from which line breaks will be removed.</param>
+        /// <returns>A new string without any line breaks.</returns>
+        public static string RemoveLineBreaks(this string str)
+        {
+            return str.Replace("\r", "").Replace("\n", "");
+        }
+
+        /// <summary>
+        /// Converts all Windows-style line endings (\r\n) to Linux-style line endings (\n) in the given string.
+        /// </summary>
+        /// <remarks>Windows形式の改行コード（\r\n）をLinux形式の改行コード（\n）に変換します。</remarks>
+        /// <param name="str">The input string with Windows-style line endings.</param>
+        /// <returns>A new string with all line endings converted to Linux-style (\n).</returns>
+        public static string ToLinuxLineEndings(this string str)
+        {
+            return str.Replace("\r\n", "\n");
+        }
+
+        /// <summary>
+        /// Converts all Linux-style line endings (\n) to Windows-style line endings (\r\n) in the given string.
+        /// </summary>
+        /// <remarks>Linux形式の改行コード（\n）をWindows形式の改行コード（\r\n）に変換します。</remarks>
+        /// <param name="str">The input string with Linux-style line endings.</param>
+        /// <returns>A new string with all line endings converted to Windows-style (\r\n).</returns>
+        public static string ToWindowsLineEndings(this string str)
+        {
+            return str.Replace("\n", "\r\n").Replace("\r\r\n", "\r\n");
+        }
+
+        /// <summary>
+        /// Writes the string content to a specified file with the specified encoding, overwriting if the file already exists.
+        /// </summary>
+        /// <remarks>エンコードを指定して指定されたファイルパスに文字列を出力します。ファイルが存在する場合は上書きされ、存在しない場合は新しく作成されます。</remarks>
+        /// <param name="str">The input string to be written to the file.</param>
+        /// <param name="filePath">The path of the file where the content will be written.</param>
+        /// <param name="encoding">The encoding to be used for writing the file.</param>
+        public static void WriteToFile(this string str, string filePath, Encoding encoding)
+        {
+            File.WriteAllText(filePath, str, encoding);
+        }
+
+        /// <summary>
+        /// Appends the string content to a specified file with the specified encoding, creating the file if it does not exist.
+        /// </summary>
+        /// <remarks>エンコードを指定して指定されたファイルパスに文字列を追加します。ファイルが存在しない場合は新しく作成されます。</remarks>
+        /// <param name="str">The input string to be appended to the file.</param>
+        /// <param name="filePath">The path of the file where the content will be appended.</param>
+        /// <param name="encoding">The encoding to be used for appending to the file.</param>
+        public static void AppendToFile(this string str, string filePath, Encoding encoding)
+        {
+            File.AppendAllText(filePath, str, encoding);
         }
     }
 }